Refactor contentHeader component to a styled component
