+++ src/locale/en.json
@@ -836,7 +836,7 @@
       "rivm": {
         "download": "",
         "href": "https://data.rivm.nl/geonetwork/srv/dut/catalog.search#/metadata/5f6bc429-1596-490e-8618-1ed8fd768427",
-        "text": "RIVM",
+        "text": "RIVM"
       }
     }
   },
@@ -875,7 +875,7 @@
       "rivm": {
         "download": "https://data.rivm.nl/covid-19/COVID-19_rioolwaterdata.csv",
         "text": "RIVM",
-        "href": "https://data.rivm.nl/geonetwork/srv/dut/catalog.search#/metadata/a2960b68-9d3f-4dc3-9485-600570cd52b9",
+        "href": "https://data.rivm.nl/geonetwork/srv/dut/catalog.search#/metadata/a2960b68-9d3f-4dc3-9485-600570cd52b9"
       }
     }
   },
